requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
pillow==11.0.0
python-docx==1.1.2
reportlab==4.2.5
typing_extensions==4.12.2
Werkzeug==3.1.3
fpdf==1.7.2

